用户数据查询

+++ src/main/java/com/cjeg/web/commen/message/Response.java
@@ -6,7 +6,7 @@ public class Response<T> {
 	//返回信息
 	private String msg;
 	//分页总数
-	private int total;
+	private long total;
 	//返回信息
 	private T obj;
 	public int getFlag() {
@@ -21,10 +21,11 @@ public class Response<T> {
 	public void setMsg(String msg) {
 		this.msg = msg;
 	}
-	public int getTotal() {
+	
+	public long getTotal() {
 		return total;
 	}
-	public void setTotal(int total) {
+	public void setTotal(long total) {
 		this.total = total;
 	}
 	public T getObj() {
